requirements.txt
alembic==1.13.1
Babel==2.15.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.5
colorama==0.4.6
docxcompose
docxtpl
et-xmlfile==1.1.0
Flask==2.3.3
Flask-Admin==1.6.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fuzzywuzzy==0.18.0
greenlet==2.0.2
gunicorn
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
Mako==1.2.4
markdown==3.4.4
MarkupSafe==2.1.3
numpy==1.25.2
openpyxl==3.0.7
pandas==2.0.3
psycopg2-binary==2.9.7
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-docx==1.0.1
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
SQLAlchemy-serializer==1.4.1
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.2.1
Werkzeug==2.3.7
WTForms==3.0.1
XlsxWriter==3.1.2
